feat(iamspanner): simplify API
All methods that read bindings are now called Read(...) and all take a
callback on the exact same format (a callback for a single binding).

No more methods on the IAMServer are called Query(...).
